sipp,  call.cpp
Route routes from the incoming messages will be converted to routes with the option as rrs="true". this option was available for only one receive message in SIPp XML file. 
Now this code has been enhanced to handle this for all the receive messages you have in the SIPp XML file. 
With this you ll be able to get the route set for each previuos receive message.